refactor: 기존의 controller를 어노테이션 기반의 컨트롤러로 수정
